clean-up group edit form
