+++ client/src/components/gatherings/EditGroup.js
@@ -1,9 +1,14 @@
 import React, { Fragment, useState, useEffect } from 'react';
+import { Link, withRouter } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import { withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
-import { TextField, IconButton } from '@material-ui/core';
-import CancelIcon from '@material-ui/icons/Cancel';
+import { FormLabel } from '@material-ui/core';
+import { FormControlLabel } from '@material-ui/core';
+import { TextField } from '@material-ui/core';
+import { Button } from '@material-ui/core';
+import { Input } from '@material-ui/core';
+import { RadioGroup } from '@material-ui/core';
+import { Radio } from '@material-ui/core';
 import { addGroup, getGroup, deleteGroup } from '../../actions/group';
 const initialState = {
     _id: '',
@@ -16,8 +21,35 @@ const initialState = {
     attendance: 0,
     notes: '',
 };
-const EditGroup = ({ addGroup, getGroup, group: { group, loading } }) => {
+
+const EditGroup = ({
+    group: { group, loading, newGroup },
+    addGroup,
+    auth: { activeRole, activeStatus },
+    getGroup,
+    match,
+    history,
+}) => {
     const [formData, setFormData] = useState(initialState);
+
+    useEffect(() => {
+        if (!group) {
+            if (match.params.gid != 0) {
+                getGroup(match.params.gid);
+            }
+        }
+        if (!loading) {
+            const groupData = { ...initialState };
+            for (const key in group) {
+                if (key in groupData) groupData[key] = group[key];
+            }
+            groupData['mid'] = match.params.mid;
+            setFormData(groupData);
+        }
+        if (match.params.gid > 0)
+            setFormData({ ...formData, groupId: match.params.gid });
+    }, [loading, getGroup, group]);
+
     const {
         _id,
         title,
@@ -29,6 +61,15 @@ const EditGroup = ({ addGroup, getGroup, group: { group, loading } }) => {
         attendance,
         notes,
     } = formData;
+
+    const handleGenderChange = (e) => {
+        console.log('btnValue:' + e.target.value);
+        setFormData({ ...formData, gender: e.target.value });
+    };
+    const handleChange = (event) => {
+        console.log('event:' + event);
+        setFormData({ ...formData, [event.target.name]: event.target.value });
+    };
     const onChange = (e) => {
         setFormData({
             ...formData,
@@ -36,23 +77,52 @@ const EditGroup = ({ addGroup, getGroup, group: { group, loading } }) => {
         });
     };
 
+    const onSubmit = (e) => {
+        e.preventDefault();
+        formData.mid = match.params.mid;
+        addGroup(formData, history, true);
+        window.scrollTo(0, 0);
+    };
     return (
         <div>
+             <form className='form' onSubmit={(e) => onSubmit(e)}>
             <header className='grpHeader'>
                 <h2>Open Share Group</h2>
             </header>
-            <select
-                key='2'
-                className='DGF-Gender'
-                name='gender'
-                value={gender}
-                onChange={(e) => onChange(e)}
-            >
-                <option value='0'>** Select Gender</option>
-                <option value='f'>Women's</option>
-                <option value='m'>Men's</option>
-                <option value='x'>Mixed</option>
-            </select>
+            <div>
+                <select
+                    key='2'
+                    // className=''
+                    name='gender'
+                    value={gender}
+                    onChange={(e) => onChange(e)}
+                >
+                    <option value='0'>** Select Gender</option>
+                    <option value='f'>Women's</option>
+                    <option value='m'>Men's</option>
+                    <option value='x'>Mixed</option>
+                </select>
+            </div>
+            <div>
+                <span>Attendance</span>
+                <span style={{ 'padding-left': 20 }}>
+                <Input
+                    // style={{ 'padding-left': 20 }}
+                    id='attendance'
+                    label='attendance'
+                    name='attendance'
+                    placeholder='0'
+                    value={attendance}
+                    type='number'
+                    size="3"
+                    maxlength='2'
+                    min='0'
+                    text-align='right'
+                    // className='attendance'
+                    onChange={(e) => onChange(e)}
+                />
+                </span>
+            </div>
             <div className='grpTitle'>
                 <TextField
                     id='title'
@@ -105,28 +175,58 @@ const EditGroup = ({ addGroup, getGroup, group: { group, loading } }) => {
                     label='Notes'
                     fullWidth
                     multiline
-                    rows='2'
+                    rows='4'
                     // variant='outlined'
                     onChange={(e) => onChange(e)}
                 />
             </div>
-            <i className={'pl-2 fas fa-check my-1'}></i>
-            <i className={'pl-3 fa fa-ban'}></i>
-            <IconButton variant='outlined' size='small' aria-label='cancel'>
-                <CancelIcon />
-            </IconButton>
+            <div className='pl-2' style={{ 'padding-top': 20 }}>
+                <input
+                    type='submit'
+                    className='btn btn-primary my-1'
+                />
+                    
+                    <span className='pl-2'>
+                    <Link
+                            className='btn btn-light my-1'
+                            to={`/editGathering/${match.params.mid}`}
+                        >
+                            Go Back
+                        </Link>
+                    </span>
+            </div>
+            </form>
         </div>
     );
+    function getGroups() {
+        // return [<div>GROUP:{match.params.gid}</div>];
+        return 'T';
+    }
+    // function giveRequestDetails() {
+    //     return [
+    //         <div>
+    //             BURP
+    //             {/* Meeting:{match.params.mid}
+    //             <br />
+    //             Group: {match.params.gid} */}
+    //         </div>,
+    //     ];
+    // }
 };
 
 EditGroup.propTypes = {
+    group: PropTypes.object.isRequired,
+    auth: PropTypes.object.isRequired,
     addGroup: PropTypes.func.isRequired,
     getGroup: PropTypes.func.isRequired,
+    deleteGroup: PropTypes.func.isRequired,
 };
+
 const mapStateToProps = (state) => ({
     group: state.group,
     auth: state.auth,
 });
-export default connect(mapStateToProps, { addGroup, getGroup })(
+
+export default connect(mapStateToProps, { addGroup, getGroup, deleteGroup })(
     withRouter(EditGroup)
 );
